Restructure some things
