+++ tests/services/cassettes/TestApiService.test_update_desk_occupied_status_is_successful.yaml
